requirements.txt
Django==2.0
django-crispy-forms
psycopg2
djangorestframework
django-allauth
django-profiles
django-filter
elasticsearch-dsl
pillow
reportlab
